Separate the PostMeta type
