+++ src/app/_types/PostMeta.tsx
@@ -1,0 +1,9 @@
+type PostMeta = {
+    slug: string;
+    title: string;
+    date: string;
+    summary: string;
+    tags: string[];
+};
+
+export default PostMeta;
